[16.0] account_chart_update - separate common data from actual tests
MT-7597
