+++ account_chart_update/tests/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
+from . import common
 from . import test_account_chart_update
